added brightness, datetime and weather modes; prepared devices for different supported things
