+++ devices/pixoo64.txt
