+++ devices/pixoomax.txt
